run: write top-level script to run debian packaging process
Write top-level script to run Debain packaging process. The script refers to
variables and functions defined in <package-name>/version.sh file to carry out
package-specific configurations and operations.

The sub-directories in the source directory are the source package names of TI's
packages. When an argument is passed to the script, it looks for a
sub-directory in the top-level directory with a matching name. Thus all
sub-directories for packages must have the same name as the source package of
that project.

Signed-off-by: Suhaas Joshi <joshiesuhaas0@gmail.com>