k3conf: Allow package to be build for any architecture
Although this tool is only useful on K3 platforms which tend to be arm64
this doesn't mean this package is arm64 only. It is a simple C program and
can build for any architecture. This could be useful with armhf when
running in AARCH32 mode. Or amd64 for testing/container builds, etc..

Signed-off-by: Andrew Davis <afd@ti.com>