run: Build and extract source packages
Currently we use the same directory to build both source and binary
packages. The downside to this is that the binary package should be built
from the source package, not a common staging directory. We also don't get
the automatic patching that happens when a source package is extracted, so
we currently have to do that manually.

Build the source package and then build the binary packages from the
extracted source.

Signed-off-by: Andrew Davis <afd@ti.com>