run: Do not package version control files into source package
This saves space in our source packages at the cost of having the source
history not available when only using the source package as the source
of the source code. :)

Version control files are often not shipped with source tarballs anyway
and links to the git repos are already available so this should be fine.

Signed-off-by: Andrew Davis <afd@ti.com>