Auto install build dependencies
Apart from run time package dependencies which are specified in the
control file of each package, to build the package itself there are some
dependencies which are not documented anywhere. These has to be found
while building by looking at the error logs. With this change, the
dependencies are installed before building the packages.

Signed-off-by: Sai Sree Kartheek Adivi <s-adivi@ti.com>