cryptodev-linux: fix postinstall scripts and add source tarball to changes file
Since the orig.tar.xz is generated everytime the package is built, the
md5sum need not exactly match from the one that got generated on a
different host or time. So add .orig.tar.xz to changes file by adding
`-sa` option to `dpkg-genchanges` command.

The existing post install script doesn't check if the file exists before
deleting the file. Which seems to be working on bookworm but is causing
a fatal error on trixie. Fix it by adding a check.

Signed-off-by: Sai Sree Kartheek Adivi <s-adivi@ti.com>